+++ CHANGELOG.md
@@ -5,11 +5,11 @@ Headlines: Added, Changed, Deprecated, Removed, Fixed, Security
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [0.6.18] - 2023-04-14
 
 ### Fixed
 
-- When `dataType = 'json'`, file data is now returned even when it has the same name as a schema field.
+- When `dataType = 'json'`, file data is now returned when it has the same name as a schema field.
 
 ### Added